benchmarking compute

+++ program/tests/functional.rs
@@ -41,6 +41,7 @@ fn setup() -> Mollusk {
     Mollusk::new(&solana_config_program::id(), "solana_config_program")
 }
 
+#[allow(clippy::arithmetic_side_effects)]
 fn get_config_space(key_len: usize) -> usize {
     let entry_size = bincode::serialized_size(&(Pubkey::default(), true)).unwrap() as usize;
     bincode::serialized_size(&(ConfigKeys::default(), MyConfig::default())).unwrap() as usize
@@ -49,7 +50,7 @@ fn get_config_space(key_len: usize) -> usize {
 
 fn create_config_account(mollusk: &Mollusk, keys: Vec<(Pubkey, bool)>) -> AccountSharedData {
     let space = get_config_space(keys.len());
-    let lamports = mollusk.sysvars.rent.minimum_balance(space as usize);
+    let lamports = mollusk.sysvars.rent.minimum_balance(space);
     AccountSharedData::new_data(
         lamports,
         &(ConfigKeys { keys }, MyConfig::default()),
@@ -66,7 +67,7 @@ fn test_process_create_ok() {
     let config_account = {
         let space = get_config_space(0);
         let lamports = mollusk.sysvars.rent.minimum_balance(space);
-        AccountSharedData::new(lamports, space as usize, &solana_config_program::id())
+        AccountSharedData::new(lamports, space, &solana_config_program::id())
     };
 
     // `instruction::initialize_account` without making it public...
@@ -518,7 +519,7 @@ fn test_config_bad_owner() {
     // Store a config account with the wrong owner.
     let config_account = {
         let space = get_config_space(keys.len());
-        let lamports = mollusk.sysvars.rent.minimum_balance(space as usize);
+        let lamports = mollusk.sysvars.rent.minimum_balance(space);
         AccountSharedData::new(lamports, 0, &Pubkey::new_unique())
     };
